+++ src/main/java/com/board/board/controller/BlogApiController.java
@@ -1,19 +1,26 @@
 package com.board.board.controller;
 
 import com.board.board.domain.Article;
-import com.board.board.dto.ArticleCreateRequest;
-import com.board.board.dto.ArticleResponse;
-import com.board.board.dto.ArticleUpdateRequest;
+import com.board.board.dto.request.ArticleCreateRequest;
+import com.board.board.dto.request.ArticleUpdateRequest;
+import com.board.board.dto.response.ArticleResponse;
 import com.board.board.service.BlogService;
 import jakarta.validation.Valid;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
 @RestController
