+++ src/main/java/com/board/board/dto/request/ArticleUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.board.board.dto;
+package com.board.board.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
