+++ src/main/java/com/board/member/controller/MemberController.java
@@ -1,9 +1,9 @@
 package com.board.member.controller;
 
-import com.board.member.dto.LoginRequest;
-import com.board.member.dto.LoginResponse;
-import com.board.member.dto.MemberSignUpRequest;
-import com.board.member.dto.MemberSignUpResponse;
+import com.board.member.dto.request.LoginRequest;
+import com.board.member.dto.request.MemberSignUpRequest;
+import com.board.member.dto.response.LoginResponse;
+import com.board.member.dto.response.MemberSignUpResponse;
 import com.board.member.service.MemberService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -29,7 +29,8 @@ public class MemberController {
     }
 
     @PostMapping("/login")
-    public ResponseEntity<LoginResponse> login(@Valid @RequestBody LoginRequest loginRequest, HttpServletResponse response) {
+    public ResponseEntity<LoginResponse> login(@Valid @RequestBody LoginRequest loginRequest,
+                                               HttpServletResponse response) {
         LoginResponse loginResponse = memberService.login(loginRequest);
 
         Cookie cookie = new Cookie("token", loginResponse.getAccessToken());
