+++ src/main/java/com/board/member/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.board.member.dto;
+package com.board.member.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
