+++ src/main/java/com/board/member/service/MemberServiceImpl.java
@@ -5,10 +5,10 @@ import com.board.exception.custom.EmailNotFoundException;
 import com.board.exception.custom.MyEntityNotFoundException;
 import com.board.exception.custom.SignUpException;
 import com.board.member.domain.Member;
-import com.board.member.dto.LoginRequest;
-import com.board.member.dto.LoginResponse;
-import com.board.member.dto.MemberSignUpRequest;
-import com.board.member.dto.MemberSignUpResponse;
+import com.board.member.dto.request.LoginRequest;
+import com.board.member.dto.request.MemberSignUpRequest;
+import com.board.member.dto.response.LoginResponse;
+import com.board.member.dto.response.MemberSignUpResponse;
 import com.board.member.entity.MemberEntity;
 import com.board.member.message.ErrorMessage;
 import com.board.member.repository.MemberRepository;
